LayoutTests/js/array-of-expected.txt

Tests for Array.of

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Array.of(1) is [1]
PASS Array.of(1, 2) is [1, 2]
PASS Array.of(1, 2, 3) is [1, 2, 3]
PASS Array.of() is []
PASS Array.of(undefined) is [undefined]
PASS Array.of('hello') is ['hello']
Construct nested Array with Array.of(Array.of(1, 2, 3), Array.of(4, 5, 6, 7, 8)).
PASS x.length is 2
PASS x[0].length is 3
PASS x[1].length is 5
Check that a setter isn't called.
PASS Array.of(1, 2, 3) did not throw exception.
PASS successfullyParsed is true

TEST COMPLETE
